internal/handler/auth_handler.go
package handler

import (
	"github.com/gin-gonic/gin"
	"hub_management_service/internal/middleware"
	"net/http"
)

// LoginRequest represents the login request body
type LoginRequest struct {
	Username string `json:"username"`
	Password string `json:"password"`
}

// LoginHandler handles login requests and issues a JWT token if credentials are valid
func LoginHandler(c *gin.Context) {
	var req LoginRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid request body"})
		return
	}

	// Hardcoded username and password validation
	if req.Username == "admin" && req.Password == "password" {
		// Generate JWT token
		token, err := middleware.GenerateJWT(req.Username)
		if err != nil {
			c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to generate token"})
			return
		}

		// Return the token
		c.JSON(http.StatusOK, gin.H{"token": token})
		return
	}

	c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid credentials"})
}
